requirements.txt
PySide6>=6.6
rdkit>=2024.3.1
SQLAlchemy>=2.0
pubchempy>=1.0.4
requests>=2.31
PyYAML>=6.0
Markdown>=3.5
Pygments>=2.17
platformdirs>=4.0
numpy>=1.26
pillow>=10.0
